+++ azdev/operations/linter/util.py
@@ -18,27 +18,29 @@ logger = get_logger(__name__)
 _LOADER_CLS_RE = re.compile('.*azure/cli/command_modules/(?P<module>[^/]*)/__init__.*')
 
 
-def filter_modules(command_loader, help_file_entries, modules=None):
+def filter_modules(command_loader, help_file_entries, modules=None, include_whl_extensions=False):
     """ Modify the command table and help entries to only include certain modules/extensions.
 
     : param command_loader: The CLICommandsLoader containing the command table to filter.
     : help_file_entries: The dict of HelpFile entries to filter.
     : modules: [str] list of module or extension names to retain.
     """
-    return _filter_mods(command_loader, help_file_entries, modules=modules)
+    return _filter_mods(command_loader, help_file_entries, modules=modules,
+                        include_whl_extensions=include_whl_extensions)
 
 
-def exclude_commands(command_loader, help_file_entries, module_exclusions):
+def exclude_commands(command_loader, help_file_entries, module_exclusions, include_whl_extensions=False):
     """ Modify the command table and help entries to exclude certain modules/extensions.
 
     : param command_loader: The CLICommandsLoader containing the command table to filter.
     : help_file_entries: The dict of HelpFile entries to filter.
     : modules: [str] list of module or extension names to remove.
     """
-    return _filter_mods(command_loader, help_file_entries, modules=module_exclusions, exclude=True)
+    return _filter_mods(command_loader, help_file_entries, modules=module_exclusions, exclude=True,
+                        include_whl_extensions=include_whl_extensions)
 
 
-def _filter_mods(command_loader, help_file_entries, modules=None, exclude=False):
+def _filter_mods(command_loader, help_file_entries, modules=None, exclude=False, include_whl_extensions=False):
     modules = modules or []
 
     # command tables and help entries must be copied to allow for seperate linter scope
@@ -48,7 +50,7 @@ def _filter_mods(command_loader, help_file_entries, modules=None, exclude=False)
     command_loader.command_table = command_table
     command_loader.command_group_table = command_group_table
     help_file_entries = help_file_entries.copy()
-    name_index = get_name_index()
+    name_index = get_name_index(include_whl_extensions=include_whl_extensions)
 
     for command_name in list(command_loader.command_table.keys()):
         try:
@@ -58,8 +60,11 @@ def _filter_mods(command_loader, help_file_entries, modules=None, exclude=False)
             logger.warning(ex)
             source_name = None
 
-        long_name = name_index[source_name]
-        is_specified = source_name in modules or long_name in modules
+        try:
+            long_name = name_index[source_name]
+            is_specified = source_name in modules or long_name in modules
+        except KeyError:
+            is_specified = False
         if is_specified == exclude:
             # brute force method of ignoring commands from a module or extension
             command_loader.command_table.pop(command_name, None)